ImageBind/requirements.txt
einops==0.8.0
ftfy==6.2.3
gradio==4.44.1
iopath==0.1.10
numpy>=1.19
Pillow==10.4.0
pydub==0.25.1
pytorchvideo==0.1.5
qdrant_client==1.11.3
regex==2024.9.11
setuptools==58.0.4
timm==0.6.7
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.5
#torch==1.13.1
#torchvision  # because torch version already specific, the right torchvision will be derived automatically
#torchaudio  # because torch version already specific, the right torchaudio will be derived automatically
#pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
#timm==0.6.7
#ftfy
#regex
#einops
#fvcore
#eva-decord==0.6.1
#iopath
#numpy>=1.19
#matplotlib
#types-regex

